+++ paper/paper.bib
@@ -10,52 +10,28 @@
 
 @article{Maertens2015,
 	Author = {Maertens, Audrey P and Weymouth, Gabriel D},
-	Date-Added = {2017-08-02 13:23:51 +0000},
-	Date-Modified = {2017-08-02 13:23:51 +0000},
+	Doi = {https://doi.org/10.1016/j.cma.2014.09.007},
 	Journal = {Computer Methods in Applied Mechanics and Engineering},
 	Pages = {106--129},
 	Publisher = {Elsevier},
 	Title = {Accurate Cartesian-grid simulations of near-body flows at intermediate Reynolds numbers},
 	Volume = {283},
-	Year = {2015},
-	Bdsk-File-1 = {YnBsaXN0MDDUAQIDBAUGJCVYJHZlcnNpb25YJG9iamVjdHNZJGFyY2hpdmVyVCR0b3ASAAGGoKgHCBMUFRYaIVUkbnVsbNMJCgsMDxJXTlMua2V5c1pOUy5vYmplY3RzViRjbGFzc6INDoACgAOiEBGABIAFgAdccmVsYXRpdmVQYXRoWWFsaWFzRGF0YV8QUS4uL0JESU0gQ01BTUUvcmV2aXNlZGNtYW1lbWFudXNjcmlwdGFuZHJlc3BvbnNlbGV0dGVycy9CRElNcGFwZXJuZXdfZGlmZl9HQUJFLnBkZtIXCxgZV05TLmRhdGFPEQJ2AAAAAAJ2AAIAAAxNYWNpbnRvc2ggSEQAAAAAAAAAAAAAAAAAAADRCJqmSCsAAAAKV/8aQkRJTXBhcGVybmV3X2RpZmZfR0FCRS5wZGYAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAxhw880nTcAAAAAAAAAAAABAAMAAAkgAAAAAAAAAAAAAAAAAAAAH3JldmlzZWRjbWFtZW1hbnVzY3JpcHRhbmQjQTU3RkYAABAACAAA0QiapgAAABEACAAAzzSdNwAAAAEAIAAKV/8ACWGoAAlgPwAJXmQACV5EAAlc3wAJQDgAApUVAAIAjU1hY2ludG9zaCBIRDpVc2VyczoAd2V5bW91dGg6AERyb3Bib3g6AERvY3VtZW50czoAUGFwZXJzOgBDRyBNZXRob2RzOgBCRElNIENNQU1FOgByZXZpc2VkY21hbWVtYW51c2NyaXB0YW5kI0E1N0ZGOgBCRElNcGFwZXJuZXdfZGlmZl9HQUJFLnBkZgAADgA2ABoAQgBEAEkATQBwAGEAcABlAHIAbgBlAHcAXwBkAGkAZgBmAF8ARwBBAEIARQAuAHAAZABmAA8AGgAMAE0AYQBjAGkAbgB0AG8AcwBoACAASABEABIAgVVzZXJzL3dleW1vdXRoL0Ryb3Bib3gvRG9jdW1lbnRzL1BhcGVycy9DRyBNZXRob2RzL0JESU0gQ01BTUUvcmV2aXNlZGNtYW1lbWFudXNjcmlwdGFuZHJlc3BvbnNlbGV0dGVycy9CRElNcGFwZXJuZXdfZGlmZl9HQUJFLnBkZgAAEwABLwAAFQACAA///wAAgAbSGxwdHlokY2xhc3NuYW1lWCRjbGFzc2VzXU5TTXV0YWJs
[…]
