### 2017-08-05 #349 @labarba
There is also no Community Guidelines, and I see that the paper mentions "engineering outreach activities and student projects," but there is no Statement of Need and no mention of research applications. You will need to work on these things, while we continue seeking commitment from a reviewer.
