### 2017-08-05 #349 @labarba
(I found a "Contributing" section in the wiki. Not where I would've expected it to be…)
