### 2017-08-05 #349 @weymouth
Is there a "best practices" reference for how to structure the github site?
I'll be more explicit on the "Statement of Need" and research application in paper.md
