### 2017-08-05 4ed9a5b
Update paper.md
