### 2017-08-05 f2b7e44
Update paper.md
