### 2017-08-12 #349 @labarba
Good question @weymouth — I don't think there is. Just emergent behavior, I suppose. Most repos will have a `CONTRIBUTING.md` file right on the root of the repo, and additionally a heading in the README that has a sentence and a link to the `md` file. Same goes for installation instructions. I think putting these things in the wiki is going to cause you some headaches, because folks won't find them and they'll ping you.
