### 2017-08-12 #349 @labarba
@weymouth : Regarding the statement of need for research applications, the paper currently has only one sentence addressing this: _"This enables users with no previous CFD experience to address open research problems in fluid mechanics; for example, how a plesusaur swims [https://doi.org/10.1017/jfm.2015.61] and how spinning cylinders can eliminate drag [https://doi.org/10.1017/jfm.2017.395]."_

First of all, you should use the `.bib` file for references and add the citations properly in the paper.

Next: mis-typed "plesiosaur."

That sentence is not sufficient to establish the research application of the software. I notice that after the mention of plesiosaur swimming as a research application, you cite the paper titled "Unsteady dynamics of rapid perching manoeuvres”—I didn't read the paper, just skimmed a few pages, but it didn't seem to be a paper on plesiosaur swimming… The next research application mentioned is drag reduction with spinning cylinders (should you use "reduce" rather than "eliminate"?), citing the paper "Flow control with rotating cylinders"—browsing that paper, it mentions Reynolds number is 47,000; but I gather that LilyPad only handles low Reynolds numbers…
