### 2017-08-14 a3b6109
Create CONTRIBUTING.md
