### 2017-08-14 d6725a6
Create INSTALL_AND_RUN.md
