+++ INSTALL_AND_RUN.md
@@ -1,0 +1,20 @@
+### Getting Started
+
+The code is written in Processing and uses Git for version control. Before working with Lily Pad you will need to get these two programs (both of these are cross platform, so get the flavour you need).
+ * Get a copy of [Processing](http://processing.org/). They have an excellent [wiki](http://wiki.processing.org/).
+ * Get a copy of [Git](http://git-scm.com/). There are many nice git [introductions](http://git-scm.com/documentation).
+
+Now you can download the Lily Pad program.  You can use the buttons to the right to [fork](https://guides.github.com/activities/forking/) the Lily Pad project, making a clone of it in your own github account, or just download the Lily Pad source as a zipfile. Alternatively, open a command terminal and type
+
+`$ git clone https://github.com/weymouth/lily-pad`
+
+Now navigate to the `lily-pad\LilyPad` folder and open `LilyPad.pde` with Processing.
+
+That's it! You're ready to run.
+
+
+### Running
+
+Once the source is opened in Processing, click the _play_ button. The default set-up (a cylinder in a free stream) will run automatically. You can drag the cylinder around in the default set-up as well. 
+
+The main Processing file is `LilyPad.pde` and this file controls what happens when you click play. Each of the files in the `LilyPad` folder has an example code section commented out at the top. By replacing the contents of the `LilyPad.pde` file with these example code sections you can run a wide variety of simulations and explore the types of geometries, flow conditions, and visualizations available.
