+++ README.md
@@ -4,8 +4,12 @@
 [![DOI](https://zenodo.org/badge/10940/weymouth/lily-pad.svg)](http://dx.doi.org/10.5281/zenodo.16065)
 
 
-* **Check out the [wiki](https://github.com/weymouth/lily-pad/wiki) to get started.**
-* Documentation on the solver lives [here](https://github.com/weymouth/lily-pad/wiki/documentation).
-* Help with [non-dimensionalization](https://github.com/weymouth/lily-pad/wiki/non-dimensional) can also be handy.
+Follow these [simple instructions to install and run Lily Pad](INSTALL_AND_RUN.md).
+
+Here is the [guideline for contributing](CONTRIBUTING) to the project.
+
+The [wiki](https://github.com/weymouth/lily-pad/wiki) has further information such as:
+* [Documentation on the solver](https://github.com/weymouth/lily-pad/wiki/documentation).
+* [Help with non-dimensionalization](https://github.com/weymouth/lily-pad/wiki/non-dimensional) can also be handy.
 
 ![LilyPad](http://weymouth.github.io/img/streakgreen2013Apr10.jpg)
