### 2017-08-14 ceee799
Update paper.md
