+++ paper/paper.bib
@@ -35,3 +35,9 @@
 	Title = {Flow Control with Rotating Cylinders},
 	Volume = {825},
 	Year = {2017}}
+
+@article{Muscutt2017,
+	Author = {L. Muscutt and G. D. Weymouth and B. Ganipathisubramani},
+	Journal = {Journal of Fluid Mechanics},
+	Title = {Performance augmentation mechanism of in-line tandem flapping foils},
+	Year = {2017}}
