### 2017-08-14 d5f9ff5
Update paper.md
