### 2017-08-14 #349 @weymouth
I have changed the organisation of the repository to follow your suggestion.

I've tried to be super clear about the need: a **simple to use** CFD solver for unsteady FSI. I can't state how important simple to use software is for actually **getting research done**! Based on the JOSS requirements, this seems to be within your remit:
 - Your software should have an obvious research application
 - Should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)

I've fixed the references (I had left out one description) and I'm waiting on the PhD student to give me the DOI for the tandem flapping foil paper. For the spinning cylinder results, the LilyPad simulations were low Re, and showed complete elimination of drag (thrust production even). LilyPad can run high Re simulations, but 2D turbulence is quite different than 3D turbulence. I have a PhD student working on that right now...
