### 2017-08-14 #349 @Kevin-Mattheus-Moerman
@labarba I am not familiar with the Processing language and my expertise is predominantly solid mechanics. However, I am interested in this submission so I will try to implement/test it. I could act as the (n+1)th reviewer, where n>1 :smile_cat:
