### 2017-08-14 7e8761e
Update paper.md
