### 2017-08-14 3e86a53
Update paper.md
