+++ paper/paper.md
@@ -34,7 +34,7 @@ bibliography: paper.bib
 
 Lily Pad was developed in 2009 to fill the need for a simple-to-use computational fluid dynamics (CFD) tool for researchers working in unsteady fluid dynamics and fluid structure interactions. Standard CFD solvers require extensive training before they can be used because hundreds of parameters, including turbulence models and numerical grid parameterizations, must be properly set before the results can be trusted. Compounding this is the fact that there is no direct feedback to quickly let users know when they have made a mistake. Open source CFD solvers such as OpenFOAM [www.openfoam.org] provide a free and adaptable alternative to standard engineering software, but do not alleviate these fundamental issues.
 
-Lily Pad addresses the need to lower the barrier to CFD in research by adopting simple high-speed methods, utilizing modern programming features and environments, and giving immediate visual feed-back to the user. Lily Pad simulates the full two-dimensional Navier-Stokes equations, but most of the complications plaguing CFD are avoided by immersing solid bodies into the fluid domain using the Boundary Data Immersion Method [@Maertens2015] on a uniform computational grid. Lily Pad is written using Processing Language [https://processing.org/] which is multi-platform and comes bundled with a development environment and for flow visualization and run-time feedback from the user from the mouse or keyboard. 
+Lily Pad addresses the need to lower the barrier to CFD in research by adopting simple high-speed methods, utilizing modern programming features and environments, and giving immediate visual feed-back to the user. Lily Pad simulates the full two-dimensional Navier-Stokes equations, but most of the complications plaguing CFD are avoided by immersing solid bodies into the fluid domain using the Boundary Data Immersion Method [@Maertens2015] on a uniform computational grid. Lily Pad is written using Processing Language [https://processing.org/] which is multi-platform and comes bundled with a development environment capable of handling run-time feedback from the user's mouse or keyboard. 
 
 Lily Pad's simple and fast usage has enabled many users with no previous CFD experience to address open research problems in fluid mechanics; for example, how a bird can produce lift and drag during perching [@Polet2015], how spinning cylinders can eliminate drag [@Schulmeister2017], and how a tandem flapping foils (such as used by a plesiosaur) can increase thrust [@Muscutt2017]. At modest resolution, the simulation speed is sufficient to run simulations in real-time and the user can interact with the solid mechanical elements to adapt the flow. This makes Lily Pad ideal for quickly testing out research concepts and use in engineering education and outreach activities [http://edition.cnn.com/videos/tv/2015/03/11/spc-mainsail-design-special-b.cnn].
 
