### 2017-08-15 02475f1
Update paper.md
