+++ paper/paper.md
@@ -36,6 +36,6 @@ Lily Pad was developed in 2009 to fill the need for a simple-to-use computationa
 
 Lily Pad addresses the need to lower the barrier to CFD in research by adopting simple high-speed methods, utilizing modern programming features and environments, and giving immediate visual feed-back to the user. Lily Pad simulates the full two-dimensional Navier-Stokes equations, but most of the complications plaguing CFD are avoided by immersing solid bodies into the fluid domain using the Boundary Data Immersion Method [@Maertens2015] on a uniform computational grid. Lily Pad is written using Processing Language [https://processing.org/] which is multi-platform and comes bundled with a development environment capable of handling run-time feedback from the user's mouse or keyboard. 
 
-Lily Pad's simple and fast usage has enabled many users with no previous CFD experience to address open research problems in fluid mechanics; for example, how a bird can produce lift and drag during perching [@Polet2015], how spinning cylinders can eliminate drag [@Schulmeister2017], and how a tandem flapping foils (such as used by a plesiosaur) can increase thrust [@Muscutt2017]. At modest resolution, the simulation speed is sufficient to run simulations in real-time and the user can interact with the solid mechanical elements to adapt the flow. This makes Lily Pad ideal for quickly testing out research concepts and use in engineering education and outreach activities [http://edition.cnn.com/videos/tv/2015/03/11/spc-mainsail-design-special-b.cnn].
+Lily Pad's simple and fast usage has enabled many users with no previous CFD experience to address open research problems in unsteady fluid mechanics; for example, how a bird can produce lift and drag during perching [@Polet2015], how spinning cylinders can eliminate drag on a bluff body [@Schulmeister2017], and how a tandem flapping foils (such as used by a plesiosaur) can increase thrust [@Muscutt2017]. At modest resolution, the simulation speed is sufficient to run simulations in real-time and the user can interact with the solid mechanical elements to adapt the flow. This makes Lily Pad ideal for quickly testing out research concepts and use in engineering education and outreach activities [http://edition.cnn.com/videos/tv/2015/03/11/spc-mainsail-design-special-b.cnn].
 
 # References
