### 2017-08-21 #349 @labarba
@weymouth : I'm still working on locating a reviewer from the Processing community. I have someone who tentatively agreed, but is unavailable for another month. Stay tuned.

Meanwhile, thank you for your patience with my pre-review enquiries on your submission.
