### 2017-08-30 dcf9bd6
Update paper.bib with new DOI
