+++ paper/paper.bib
@@ -38,6 +38,7 @@
 
 @article{Muscutt2017,
 	Author = {L. Muscutt and G. D. Weymouth and B. Ganipathisubramani},
+	Doi = {https://doi.org/10.1017/jfm.2017.457},
 	Journal = {Journal of Fluid Mechanics},
 	Title = {Performance augmentation mechanism of in-line tandem flapping foils},
 	Year = {2017}}
