### 2017-08-30 #349 @weymouth
I added the DOI for the JFM plesiosaur paper (https://doi.org/10.1017/jfm.2017.457) to the paper.bib.

Our Royal Society plesiosaur paper also just came out (https://t.co/2iUZsQPyzA) if you're interested.
