### 2017-09-16 #394 @alignedleft
## General Comments

*This project is impressive.* Processing is a great tool for building real-time, interactive tools. And the language is designed for beginners (in as much as possible), so I appreciate the choice of Processing as the language used here—to reach a less technical (in terms of coding skills) audience.

I have no expertise in fluid dynamics, however, so I can't comment on the value of the specific modeling techniques used here.  It seems reasonable to concur that LilyPad is making these approaches much more accessible to people in this field (students and otherwise).

All of that said, the Processing IDE presents some challenges for a project of this scope. Processing philosophy of coding is that of "sketching in code," and Processing "programs" are called "sketches." Most sketches consist of a single .pde file, as in this included example:

![screen shot 2017-09-15 at 6 41 56 pm](https://user-images.githubusercontent.com/1034002/30508193-133b403e-9a46-11e7-92a2-85bee4ef67fd.png)

However, the IDE does support splitting your code across multiple .pde files. This is typically used to store one class per file, as in this other included example:

![screen shot 2017-09-15 at 6 42 00 pm](https://user-images.githubusercontent.com/1034002/30508198-216eee6c-9a46-11e7-8773-de28c425a590.png)

Note that each .pde file is rendered as a separate tab in the IDE.

The IDE is not intended for advanced use cases, such as having 30+ files per sketch, as LilyPad does:

![screen shot 2017-09-15 at 6 38 33 pm](https://user-images.githubusercontent.com/1034002/30508162-1b0d4c04-9a45-11e7-938f-aecf8e4c226a.png)

Unfortunately, with so many tabs squished onto the screen, tab titles are obscured, which makes the practical task of navigating the examples challenging.  (The workaround is to use the downward arrow at far right.)  This is not exactly a critique of LilyPad, but normally the Processing team [recommends that advanced users use a third-party IDE like Eclipse](https://github.com/processing/processing-tool-template).  Yet this project is trying to reach a new-to-Processing audience.

Given that, I would recommend repackaging the examples as separate sketches, each in its own folder.  Yes, this would probably be more work for the maintainer (and I see how having everything in one monster-sized sketch is more convenient, for that purpose!), but it would be much more accessible to others.  If creating separate sketch files sounds like a headache, you could imagine using custom build scripts to, say, extract the relevant pieces from the main LilyPad directory, and repackage each example as its own sketch.

Doing this would avoid the confusion (and errors) that occur when copy-pasting commented-out code into the main LilyPad.pde tab.  I found this copy-and-paste instruction unclear, and it took some trial and error to figure out which pieces to copy and where to paste them (which, itself, required overwriting some existing code).

Having separate sketches is "the Processing way" to package code. The Processing team has [documentation on how to contribute your own examples for the built-in Examples browser](https://github.com/processing/processing/wiki/Examples-Overview).  Simply to make this project more readily accessible by a wider audience, I'd recommend repackaging as described above.

## Review Checklist

I have a few comments about the unchecked items in the checklist above.

### Documentation

**Example usage:** All of the examples are in the code itself, and while the example do contain some inline comments, I would find it extremely beneficial for the primary documentation (on the project wiki) to step through a few of the simplest examples, line by line, and explain what's going on.  This may be especially important since the intended audience is new to Processing; they will need clarity on what parts are Processing-specific syntax, and what parts are about the modeling techniques being used.

**Functionality documentation:** The documentation describes the primary classes used quite generally, but each class’s API is not documented formally.  I would expect a list of all fields and methods (including arguments for each, and returned values).  Without that documentation, it is a great deal more effort for users to customize the software to their needs.

**Automated tests:** There is no documentation of this, but, to be fair, Processing is not really set up for automated testing, in the usual sense. When the docs mention letting people "test" new submissions, it could be useful to clarify exactly what's meant here.

**Community guidelines:** These could also be clarified, especially for this new-to-Processing audience.  The docs don't mention how to report problems (perhaps GitHub issues are assumed) or seek support.

### Software paper

**References:** No references were provided.

## Conclusion

Wonderful project, and a great use of Processing. All of my comments are purely about packaging and documentation, to help this tool reach the widest possible audience—or to help the existing audience employ it more rapidly and successfully.
