### 2017-10-01 #394 @labarba
@weymouth 👋

The reviewer had some requests for improvements. When are you able to address these?
