### 2017-10-02 #394 @weymouth
I am still trying to figure out if the subfolder idea of @alignedleft can be done. It is simple enough to move items into folders. The trick is to pull this into a new sketch when the users want somehow.

>  If creating separate sketch files sounds like a headache, you could imagine using custom build scripts to, say, extract the relevant pieces from the main LilyPad directory, and repackage each example as its own sketch.

I can see the appeal of chunking the many files into smaller, more manageable pieces, I'm not sure how to realize this suggestion. Any further ideas from @alignedleft are welcome...

I'm working on transferring information into an example walkthrough and functional documentation.
