### 2017-10-02 #394 @alignedleft
Sure! I wasn't imagining that you could move items around in folders dynamically. Rather, you could run a script locally, to pre-package the scripts into discrete sketches that could each be run independently — in contrast to the current, single, multi-tabbed sketch, which requires manual copy/pasting.

The "script" I'm proposing wouldn't / couldn't be written in Processing (I don't think). Maybe Python or similar would be more appropriate. The script would just be a convenience (for you).  Think of it as doing all that copy-pasting for your audience, in advance.

We use lots of various build scripts for the processing.org website, not to accomplish this exact same task, but to repackage the included Examples in various forms—e.g., this example is packaged with the Processing download itself, but is also repackaged and formatted for display on the website:

https://processing.org/examples/flocking.html
