### 2017-10-02 #394 @weymouth
Got it. Of course, this means the core code (10+ pde files) would need to be copied into each of those example folders. That's why you were worried about maintenance, right? Is there any way to avoid this? Dynamic aliases or something?

Otherwise I guess the development model would be to check if any of those repeated files had been changed before pushing new versions to git. They could also use this to test any changes against each of the example cases.

Do you think that will significantly improve the usage case for research users? It seems like it might get in the way of extending/combining the current examples...
