### 2017-10-02 #394 @alignedleft
@weymouth I am not really the target audience, since fluid dynamics is not my field.  :-)  But generally speaking, my thought was simply that it would be easier for new users if they could simply open an example and click "run" to immediately see results.  (That's "the Processing way.")

Even if people are really committed to using the software in advance (such as because, say, they are grad students and have been assigned the software as homework or for a project), having the examples nicely packages will help them understand the breadth of what is possible using this tool.  Currently, it takes a lot of fussy and careful copy and pasting to explore that.  Even for someone familiar with Processing, it may be unclear why the copy/pasting is necessary.

If you don't feel it's worth it to repackage the examples, you might consider making a demo video (like a showreel) that showcases each of the examples briefly in turn.
