### 2017-12-10 #394 @weymouth
Sorry for dropping off the radar for a while. I tried (and failed) twice to incorporate the suggested organizational changes to the project. Both attempts significantly hampered the development workflow for anything other than the "main trunk" of the project, and I ended up having to revert back to the current organization and redo the development.

I gave this a solid try since I can completely see the reviewer's point. It would be great to have these examples be plug and play. However, this development style just doesn't seem to be what the processing IDE was set up for. Everything pretty much needs to be in the one sketch or things gets very complicated very quickly.

So, my plan is to clean up the other points raised and make a Youtube tutorial which I'll link to from github.
