### 2017-12-12 #394 @alignedleft
Sounds like a great compromise. My main goal in my comments was to help increase accessibility of the examples to beginner users. Since Processing is organized around individual sketches, it makes sense to do as much as you can within that limitation. Making a tutorial video and also just using consistent formatting within each tab will help, too. For example, use some sort of unique comment lines to indicate where the user should copy and paste, e.g.:

```
// *^*^*^*^*^*^*^*^*^*^*^*^*^*^*^*
// *^*^*^*^*^*^*^*^*^*^*^*^*^*^*^*

// …example code here…

// *^*^*^*^*^*^*^*^*^*^*^*^*^*^*^*
// *^*^*^*^*^*^*^*^*^*^*^*^*^*^*^*
```
