### 2018-06-15 #394 @arfon
I've just send @weymouth an email letting him know that if we don't hear back from him in the next couple of weeks we'll assume he no longer wishes to publish this paper.
